agents/Day-23-Job-Search-Assistant/career-assistant/pyproject.toml
[project]
name = "career-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "pydantic>=2.11.5",
    "pydantic-ai>=0.2.12",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
]
